style: forge fmt + named imports

+++ src/Counter.sol
@@ -2,13 +2,13 @@
 pragma solidity ^0.8.13;
 
 contract Counter {
-    uint256 public number;
+  uint public number;
 
-    function setNumber(uint256 newNumber) public {
-        number = newNumber;
-    }
+  function setNumber(uint newNumber) public {
+    number = newNumber;
+  }
 
-    function increment() public {
-        number++;
-    }
+  function increment() public {
+    number++;
+  }
 }
