+++ test/Counter.t.sol
@@ -2,23 +2,23 @@
 pragma solidity ^0.8.13;
 
 import "forge-std/Test.sol";
-import "../src/Counter.sol";
+import {Counter} from "src/Counter.sol";
 
 contract CounterTest is Test {
-    Counter public counter;
+  Counter public counter;
 
-    function setUp() public {
-        counter = new Counter();
-        counter.setNumber(0);
-    }
+  function setUp() public {
+    counter = new Counter();
+    counter.setNumber(0);
+  }
 
-    function testIncrement() public {
-        counter.increment();
-        assertEq(counter.number(), 1);
-    }
+  function testIncrement() public {
+    counter.increment();
+    assertEq(counter.number(), 1);
+  }
 
-    function testSetNumber(uint256 x) public {
-        counter.setNumber(x);
-        assertEq(counter.number(), x);
-    }
+  function testSetNumber(uint x) public {
+    counter.setNumber(x);
+    assertEq(counter.number(), x);
+  }
 }
